backend/pyproject.toml
[project]
name = "image-optimizer-backend"
version = "0.1.0"
description = "Internal image optimization tool - FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.19",
    "pillow>=11.1.0",
    "pillow-avif-plugin>=1.4.6",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
]
